% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_json_str.R
\name{cat_json_str}
\alias{cat_json_str}
\title{Concatenate stringified JSON}
\usage{
cat_json_str(txt)
}
\arguments{
\item{txt}{character: the stringified JSON to be concatenated.}
}
\value{
character: a string (stringified JSON), stripped of newlines and
  spaces following colons or commas.
}
\description{
Strips 'stringified' JSON of newlines and spaces following colons or commas.
}
\examples{
cat_json_str("{\n    'v1': [20],\n    'v2': [1.5]\n  }")

}
