% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_utils.R
\name{compact_json_str}
\alias{compact_json_str}
\title{Compact stringified JSON}
\usage{
compact_json_str(txt)
}
\arguments{
\item{txt}{character: the stringified JSON to be compacted.}
}
\value{
character: a string (stringified JSON), stripped of newlines and
  spaces following colons or commas.
}
\description{
Strips 'stringified' JSON of newlines and spaces following colons or commas.
}
\examples{
compact_json_str("{\\n    'v1': [20],\\n    'v2': [1.5]\\n  }")

}
