% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_utils.R
\name{fmt_int}
\alias{fmt_int}
\title{Convert integers into strings}
\usage{
fmt_int(num)
}
\arguments{
\item{num}{integer vector of values to be converted (these do not need to be
of class \code{integer} but they do need to \emph{be} integer, i.e. without
any digits after the decimal points).}
}
\value{
character: vector of formatted integer values, padded to the length
  of the largest value with leading spaces.
}
\description{
Converts an integer value or vector of integer values into string(s) without
any decimal point, padding them with leading zeroes to ensure they are all
the same length.
}
\examples{
fmt_int(1.0)
fmt_int(2)
fmt_int(c(2, 25, 467))

}
