% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_utils.R
\name{get_incremental_names}
\alias{get_incremental_names}
\title{Construct sequence of incremental names}
\usage{
get_incremental_names(n_items, prefix = NULL, sep = "_")
}
\arguments{
\item{n_items}{integer: the number of names to generate.}

\item{prefix}{character: an optional name prefix to use. E.g. if
\code{prefix} is "xyz" and \code{n_items} is 3 this will result in names
\code{xyz_1}, \code{xyz_2} and \code{xyz_3}. If NULL the returned names
will simply be a sequence of integers which are formatted as strings.}

\item{sep}{character: separator to use between \code{prefix} and the
incremental numbers (default is an underscore). Ignored if \code{prefix =
NULL}.}
}
\value{
a character vector of names with successively incremented integer
  suffixes.
}
\description{
Constructs a sequence of names, each of which has an increasing integer
appended to it.
}
\examples{
get_incremental_names(2, "item")
get_incremental_names(3, "item", sep = "")
get_incremental_names(11, "z")
get_incremental_names(12)

}
