% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_utils.R
\name{get_max_int_len}
\alias{get_max_int_len}
\title{Compute maximum length of formatted integers}
\usage{
get_max_int_len(num)
}
\arguments{
\item{num}{integer: vector of the integer values.}
}
\value{
integer: maximum character length of the integers provided, if they
  were formatted as strings.
}
\description{
Computes the maximum character length of a vector of integers when they are
converted to strings.
}
\examples{
get_max_int_len(1.0)
get_max_int_len(2)
get_max_int_len(c(2, 25, 467))

}
