% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_seed_val.R
\name{get_random_seed_val}
\alias{get_random_seed_val}
\title{Generate random seed value}
\usage{
get_random_seed_val()
}
\value{
integer: a random value (sampled from the numbers between 1 and
.Machine$integer.max).
}
\description{
Generates a random seed value for setting the RNG seed by sampling a random
number (between 1 and .Machine$integer.max).
}
\examples{
get_random_seed_val()

}
\seealso{
\code{\link{set_init_seed}}
}
