% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_utils.R
\name{get_rc_fld_vals}
\alias{get_rc_fld_vals}
\title{Get field values in an RC object (with optional replacement)}
\usage{
get_rc_fld_vals(rc_obj, vals = NULL)
}
\arguments{
\item{rc_obj}{an RC object.}

\item{vals}{list: fields to replace in \code{rc_obj}.}
}
\value{
list: existing fields in \code{rc_obj} with values replaced by any
  field values supplied in \code{vals}.
}
\description{
Retrieves a list of field values from an RC object, replacing any existing
field values with those supplied.
}
\examples{
thing_factory <- setRefClass(
  "thing", fields = list(name  = "character", value = "numeric")
)
flibgib <- thing_factory$new(name = "Flibberty Gibbet", value = 1e6)
get_rc_fld_vals(flibgib)
get_rc_fld_vals(flibgib, list(value = 25.6))

}
