% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_invertible.R
\name{is_invertible}
\alias{is_invertible}
\title{Check if a matrix is invertible}
\usage{
is_invertible(x)
}
\arguments{
\item{x}{matrix: numeric matrix which should be checked to see if it is
invertible.}
}
\value{
logical: TRUE if the supplied matrix is invertible, FALSE otherwise.
}
\description{
Checks a matrix to see if it is invertible.
}
\examples{
is_invertible(matrix(c(1, 2, 2, 1), 2, 2))  # => TRUE

is_invertible(matrix(c(1, 1, 1, 1), 2, 2))  # => FALSE

}
