% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sym_pos_def.R
\name{is_sym_pos_def}
\alias{is_sym_pos_def}
\title{Check if a matrix is symmetric positive-definite}
\usage{
is_sym_pos_def(x)
}
\arguments{
\item{x}{matrix: numeric matrix which should be checked to see if it is
symmetric positive-definite.}
}
\value{
logical: TRUE if the supplied matrix is symmetric positive-definite,
FALSE otherwise.
}
\description{
Checks a matrix to determine if it is symmetric positive-definite.
}
\examples{
is_sym_pos_def(matrix(c(2.51, 2.01, 2.01, 1.74), 2, 2))  # => TRUE
is_sym_pos_def(matrix(c(1, 1, 1, 1), 2, 2))              # => FALSE

}
