% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_utils.R
\name{lgl_str}
\alias{lgl_str}
\title{Convert logical variable to string}
\usage{
lgl_str(lgl, case = NULL)
}
\arguments{
\item{lgl}{logical: the value to be converted to a string.}

\item{case}{character: case of the output ("title", "lower" or "upper").}
}
\value{
character: the logical value as a string, i.e. "yes" or "no",
  formatted according to the specified \code{case} option.
}
\description{
Converts a logical variable into a string representation.
}
\examples{
lgl_str(FALSE)                 # => "No"
lgl_str(TRUE, case = "upper")  # => "YES"
lgl_str(FALSE, case = "lower") # => "no"

}
