% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_or.R
\name{logical_OR}
\alias{logical_OR}
\alias{\%||\%}
\title{Logical \code{OR} operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{NULL, or a particular value.}

\item{b}{the default value to return if the first operand is NULL.}
}
\value{
a non-NULL value. This is is the value of the first operand if it is
  non-NULL, or the value of the second operand if the first operand is NULL.
}
\description{
Sets default values using a logical \code{OR} operator.
}
\details{
This infix function is inspired by Ruby's "logical OR" operator (\code{||})
for conveniently setting default values.\cr

usage:\cr
\code{function_that_might_return_null() \%||\% <default value>}

See Advanced R (Hadley Wickham) Section 6.5.1 p. 91.
}
\examples{
x <- NULL
x \%||\% 10              # => 10

x <- "Madiba"
y <- "Nelson Mandela"
x \%||\% y               # => 'Madiba'

}
\references{
Wickham, Hadley (2015). Advanced R, CRC Press, Boca Raton FL.
}
