% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{match_col_class}
\alias{match_col_class}
\title{Get column indexes of a given class in a data frame}
\usage{
match_col_class(df, class_type)
}
\arguments{
\item{df}{data.frame: data frame for which columns should be matched.}

\item{class_type}{character: type of class to match (e.g. "numeric",
"integer", "logical", "character", "factor").}
}
\value{
integer: the vector of column indexes whose classes match the
  specified class.
}
\description{
For a specified class such as "numeric" or "character", gets the column
indexes in a data frame for columns matching the class.
}
\examples{
id    <- letters[1:3]
size  <- c(1.2, 3.5, 4.1)
num   <- c(1, 2, 3)
intgr <- c(1L, 2L, 3L)

df_no_fac <- data.frame(id, size, num, intgr, stringsAsFactors = FALSE)
match_col_class(df_no_fac, "character")

df_with_fac <- data.frame(id, size, num, intgr)
match_col_class(df_with_fac, "character")
match_col_class(df_with_fac, "factor")
match_col_class(df_with_fac, "numeric")
match_col_class(df_with_fac, "integer")

}
