% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_mtx_from_vecs.R
\name{mk_mtx_from_vecs}
\alias{mk_mtx_from_vecs}
\title{Create matrix from vectors}
\usage{
mk_mtx_from_vecs(...)
}
\arguments{
\item{...}{one or more named vectors which should be combined into the
columns of a matrix. If the vectors are named, these names will become the
column names. Note that the vectors should all be of the same length and
type.}
}
\value{
a matrix constructed from the named vectors.
}
\description{
Binds the supplied vectors into columns of a matrix consisting with vector
names as column names (if the vectors are named).
}
\examples{
mk_mtx_from_vecs(col_1 = c("a", "b"), col_2 = c("alpha", "bravo"))

mk_mtx_from_vecs(A = c(1, 10, 100))

mk_mtx_from_vecs(seq(10, 30, 10), seq(70, 90, 10))

}
