% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{notin}
\alias{notin}
\alias{\%notin\%}
\title{Find non-matching values}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector or NULL: values to be compared.}

\item{y}{vector or NULL: values for which \emph{non}-matches should be
returned.}
}
\value{
a logical vector indicating if there is \emph{no} match for its left
  operand.
}
\description{
Returns non-matching values for its left operand when compared to the right
operand.
}
\examples{
foo <- letters[1:6]
foo[foo \%notin\% c("a", "c", "e")]
# => [1] "b" "d" "f"

lst <- list(a = 1, b = 2, c = 3, d = 4)
lst[names(lst) \%notin\% c("a", "c", "d")]
# =>
# $b
# [1] 2

NULL \%notin\% c(1, 2)
# => logical(0)

c("a", "b", "c") \%notin\% NULL
# => [1] TRUE TRUE TRUE

}
