% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_c.R
\name{pad_c}
\alias{pad_c}
\title{Pad and join multiple strings within character vectors}
\usage{
pad_c(..., pad_last = FALSE, sep = " ", collapse = NULL)
}
\arguments{
\item{...}{one or more character vectors as for \code{\link[stringr]{str_c}}.}

\item{pad_last}{logical: pad the last vector in addition to the others or
leave it without padding?}

\item{sep}{character: separator between vector elements.}

\item{collapse}{character: optional string used to combine input vectors into
a single string.}
}
\value{
character: a single string constructed in a similar way as with
  \code{\link[stringr]{str_c}} but where vectors are padded to a uniform
  width. If collapse is NULL (the default) the returned value is a character
  vector with length equal to the length of the first input vector. If
  collapse is non-NULL, the returned value is a character vector of length 1.
}
\description{
Joins one or more character vectors as in \code{\link[stringr]{str_c}} but
with the character strings in each vector padded to make them equal length.
The last vector is by default left unpadded but can optionally be padded too.
}
\examples{
vec1 <- c("Avocado:", "Blueberry:", "Coconut:", "Dill:")
vec2 <- c(rep("fruit", 3), "herb")
vec3 <- c("green", "purple", "white", "green")

pad_c(vec1, pad_last = TRUE)

cat(pad_c(vec1, vec2, collapse = "\n"), "\n")

cat(pad_c(vec1, vec2, vec3, collapse = "\n"), "\n")

}
\seealso{
\code{\link[stringr]{str_c}}
}
