% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl}
\alias{pl}
\title{Pluralise word}
\usage{
pl(singular, num, irregular = NULL)
}
\arguments{
\item{singular}{character: singular form of the word.}

\item{num}{integer: number of items. \code{num = 1} will give the singular
form; \code{num = 0} or \code{num > 1} will give the plural form.}

\item{irregular}{character: plural form of the word if it is irregular and
needs to be explicitly supplied.}
}
\value{
character: singular or plural form of the word, depending on the
  value of \code{num}.
}
\description{
Converts a word into a plural (English).
}
\details{
This function would usually be used in a context where the number of items
  \code{num} is a variable and not known at run-time.
}
\examples{
pl("shoe", 2)
pl("shoe", 1)

n <- 0
pl("thing", n)

pl("sheep", 5)           # => incorrect (gives "sheeps")
pl("sheep", 5, "sheep")  # => correct   (gives "sheep")

}
