% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rng_utils.R
\name{set_init_seed}
\alias{set_init_seed}
\title{Set RNG seed}
\usage{
set_init_seed(seed_val)
}
\arguments{
\item{seed_val}{one of the following:\itemize{
\item{a single character value (which can be interpreted as an integer) to
set the RNG;}
\item{the character literal "random" (to set a random seed);}
\item{\code{NULL}, to re-initialize the RNG as if no seed had yet been set
(see 'Details' in \code{\link{set.seed}}); or}
\item{\code{NA}, in which case the function will return without any change
to the seed.}
}}
}
\value{
NULL, invisibly
}
\description{
Sets the RNG (random number generator) seed.
}
\examples{
\dontrun{
set_init_seed(1)        # => set seed to a specific value
set_init_seed("random") # => set seed to a randomly-generated value
set_init_seed(NULL)     # => re-initialize seed as if none had yet been set
set_init_seed(NA)       # => return immediately without changing seed
}

}
\seealso{
\code{\link{set.seed}}, \code{\link{get_random_seed_val}}
}
\keyword{RNG}
\keyword{random}
\keyword{seed,}
