% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_lst_lst.R
\name{simplify_lst_lst}
\alias{simplify_lst_lst}
\title{Simplify list of lists}
\usage{
simplify_lst_lst(lst, unnamed_prefix = "name")
}
\arguments{
\item{lst}{a list of lists, each of which should be moved up one level in the
list hierarchy.}

\item{unnamed_prefix}{character: string to use as a prefix for naming unnamed
list items in the event that any list items are unnamed.}
}
\value{
list: the items in \code{lst} but each moved one level up in the list
  hierarchy, resulting in a list of single elements.
}
\description{
Simplify a list of lists by eliminating the top-level list.
}
\details{
List names are handled as follows: if any list items are named, these names
will be set in the returned list. If any list items are unnamed, their names
will be set to the default value prefixed by \code{unnamed_prefix} with an
integer suffix starting at 1 and incremented for each unnamed item.
}
\examples{
a_1 <- "pineapple"
a_2 <- list(uno = 1, due = 2, tre = 3)
a_3 <- c(1, 4, 9, 16, 25)
lst <- list(list(z_1 = a_1), list(z_2 = a_2), list(z_3 = a_3))
simplify_lst_lst(lst)

a_1 <- "pineapple"
a_2 <- list(uno = 1, due = 2, tre = 3)
a_3 <- c(1, 4, 9, 16, 25)
lst <- list(list(a_1), list(z = a_2), list(a_3))
simplify_lst_lst(lst)

}
