% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringify_vec.R
\name{stringify_vec}
\alias{stringify_vec}
\title{Stringify vector}
\usage{
stringify_vec(vec, item_sep = ", ", name_sep = " = ", fmt = NULL)
}
\arguments{
\item{vec}{vector: the vector to be converted into a character string.}

\item{item_sep}{character: separator to use between vector elements when
converting the vector into a string.}

\item{name_sep}{character: separator to use between names and values when
converting the vector into a string (if names exist).}

\item{fmt}{character: format specification for the vector values, or NULL to
use defaults (which would be as returned by \code{\link{format}}).}
}
\value{
character: a single character string representation of the vector.
}
\description{
Converts a vector (named or unnamed) into a single string representation.
}
\examples{
stringify_vec(c(1, 2, 3))

stringify_vec(c(one = 1, two = 2, three = 3))

}
