% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titleize_first.R
\name{titleize_first}
\alias{titleize_first}
\title{Make first word title case}
\usage{
titleize_first(txt)
}
\arguments{
\item{txt}{character: text to be formatted in title case.}
}
\value{
character: the supplied text with its first word in title case.
}
\description{
Formats the first word in the given text to have title case.
}
\examples{
titleize_first("hello")

titleize_first("hello world")

}
